+++ frontend/src/app/components/config/webui_config_view.rs
@@ -1,8 +1,8 @@
-use crate::app::components::Card;
+use crate::app::components::{Card, Chip};
 use crate::app::context::ConfigContext;
 use crate::{
-    config_field, config_field_bool, config_field_bool_empty, config_field_empty,
-    config_field_hide, config_field_optional,
+    config_field, config_field_bool, config_field_bool_empty, config_field_child,
+    config_field_empty, config_field_hide, config_field_optional,
 };
 use yew::prelude::*;
 use yew_i18n::use_translation;
@@ -16,8 +16,8 @@ const LABEL_USERFILE: &str = "LABEL.USERFILE";
 const LABEL_PLAYER_SERVER: &str = "LABEL.PLAYER_SERVER";
 const LABEL_USER_UI_ENABLED: &str = "LABEL.USER_UI_ENABLED";
 const LABEL_CONTENT_SECURITY_POLICY: &str = "LABEL.CONTENT_SECURITY_POLICY";
+const LABEL_CONTENT_SECURITY_POLICY_CUSTOM_ATTRIBUTES: &str = "LABEL.CUSTOM_ATTRIBUTES";
 const LABEL_PATH: &str = "LABEL.PATH";
-
 
 #[function_component]
 pub fn WebUiConfigView() -> Html {
@@ -42,11 +42,18 @@ pub fn WebUiConfigView() -> Html {
            <>
             { config_field_bool_empty!(translate.t(LABEL_ENABLED)) }
             { config_field_bool_empty!(translate.t(LABEL_USER_UI_ENABLED)) }
-            { config_field_bool_empty!(translate.t(LABEL_CONTENT_SECURITY_POLICY)) }
+            { config_field_child!(translate.t(LABEL_CONTENT_SECURITY_POLICY), {
+                html! {
+                    <>
+                        { config_field_bool_empty!(translate.t(LABEL_ENABLED)) }
+                        { config_field_empty!(translate.t(LABEL_CONTENT_SECURITY_POLICY_CUSTOM_ATTRIBUTES)) }
+                    </>
+                }
+            }) }
             { config_field_empty!(translate.t(LABEL_PATH)) }
             { config_field_empty!(translate.t(LABEL_PLAYER_SERVER)) }
-            { render_empty_auth()}
-            </>
+            { render_empty_auth() }
+           </>
         }
     };
 
@@ -60,7 +67,36 @@ pub fn WebUiConfigView() -> Html {
                         <>
                             { config_field_bool!(web_ui, translate.t(LABEL_ENABLED), enabled) }
                             { config_field_bool!(web_ui, translate.t(LABEL_USER_UI_ENABLED), user_ui_enabled) }
-                            { config_field_bool!(web_ui, translate.t(LABEL_CONTENT_SECURITY_POLICY), content_security_policy) }
+                            { config_field_child!(translate.t(LABEL_CONTENT_SECURITY_POLICY), {
+                                html! {
+                                    match web_ui.content_security_policy.as_ref() {
+                                        Some(csp) => html! {
+                                            <>
+                                                { config_field_bool!(csp, translate.t(LABEL_ENABLED), enabled) }
+                                                { config_field_child!(translate.t(LABEL_CONTENT_SECURITY_POLICY_CUSTOM_ATTRIBUTES), {
+                                                    html! {
+                                                        <div class="tp__config-view__tags">
+                                                            {
+                                                                if let Some(custom) = &csp.custom_attributes {
+                                                                    html! { for custom.iter().map(|a| html! { <Chip label={a.clone()} /> }) }
+                                                                } else {
+                                                                    html! {}
+                                                                }
+                                                            }
+                                                        </div>
+                                                    }
+                                                }) }
+                                            </>
+                                        },
+                                        None => html! {
+                                            <>
+                                                { config_field_bool_empty!(translate.t(LABEL_ENABLED)) }
+                                                { config_field_empty!(translate.t(LABEL_CONTENT_SECURITY_POLICY_CUSTOM_ATTRIBUTES)) }
+                                            </>
+                                        }
+                                    }
+                                }
+                            }) }
                             { config_field_optional!(web_ui, translate.t(LABEL_PATH), path) }
                             { config_field_optional!(web_ui, translate.t(LABEL_PLAYER_SERVER), player_server) }
                             <Card>